First step of refactoring move file into an input package
Uses type aliases and variable aliases to move all the execution into
the input package
