+++ filebeat/input/prospector_test.go
@@ -1,3 +1,3 @@
 // +build !integration
 
-package prospector
+package input
