+++ filebeat/input/registry_test.go
@@ -1,4 +1,4 @@
-package prospector
+package input
 
 import (
 	"testing"
